+++ tests/testthat/test-footywire.R
@@ -1,5 +1,9 @@
 context("testing footywire connections")
 
+if (!testthat:::on_cran()) {
+  fw_dat <- update_footywire_stats()
+}
+
 
 test_that("get_footywire_stats work with different inputs", {
   testthat::skip_if_offline()
@@ -194,12 +198,24 @@ describe("get_footywire_betting_odds", {
 test_that("update_footywire_stats works ", {
   testthat::skip_if_offline()
   testthat::skip_on_cran()
-  
-fw_dat <- update_footywire_stats()
 
 expect_type(fw_dat, "list")
 #expect_equal(fw_dat[1,1], "2010-03-25")
 expect_error(update_footywire_stats("a"))
 
+})
+
+test_that("no duplicate games in footywire data,", {
+  testthat::skip_if_offline()
+  testthat::skip_on_cran()
+  
+  n_duplicates <- fw_dat %>%
+    dplyr::ungroup() %>%
+    dplyr::group_by(Date, Season, Round, Team, Player) %>%
+    dplyr::summarise(count_rows = dplyr::n()) 
+  
+  expect_lte(max(n_duplicates$count_rows), 1)
+  
+  
 })