+++ NEWS.md
@@ -1,3 +1,5 @@
+# fitzRoy (development version)
+
 # fitzRoy 0.3.1
 
 * Updating vignettes to use internal data rather than downloading from the internet